+++ rust/lance-encoding/Cargo.toml
@@ -39,6 +39,7 @@ rand_xoshiro = "0.6.0"
 hyperloglogplus.workspace = true
 
 [dev-dependencies]
+lance-testing.workspace = true
 rand.workspace = true
 tempfile.workspace = true
 test-log.workspace = true
