+++ rust/lance-encoding/src/encodings/physical/value.rs
@@ -314,6 +314,7 @@ pub(crate) mod tests {
 
     use lance_arrow::DataTypeExt;
     use lance_datagen::{array::rand_with_distribution, ArrayGenerator};
+    use lance_testing::util::EnvVarGuard;
 
     use crate::{
         encoder::{ArrayEncoder, CoreBufferEncodingStrategy},
@@ -359,7 +360,7 @@ pub(crate) mod tests {
 
     #[test_log::test(test)]
     fn test_will_bitpack_allowed_types_when_possible() {
-        std::env::set_var("LANCE_USE_BITPACKING", "TRUE");
+        let _env_guard = EnvVarGuard::new("LANCE_USE_BITPACKING", "TRUE");
         let test_cases: Vec<(DataType, ArrayRef, u64)> = vec![
             (
                 DataType::UInt8,
@@ -467,7 +468,6 @@ pub(crate) mod tests {
                 }
             }
         }
-        std::env::remove_var("LANCE_USE_BITPACKING");
     }
 
     struct DistributionArrayGeneratorProvider<
@@ -519,7 +519,7 @@ pub(crate) mod tests {
 
     #[test_log::test(tokio::test)]
     async fn test_bitpack_primitive() {
-        std::env::set_var("LANCE_USE_BITPACKING", "TRUE");
+        let _env_guard = EnvVarGuard::new("LANCE_USE_BITPACKING", "TRUE");
         let bitpacked_test_cases: &Vec<(DataType, Box<dyn ArrayGeneratorProvider>)> = &vec![
             // check less than one byte for multi-byte type
             (
@@ -611,7 +611,5 @@ pub(crate) mod tests {
             let field = Field::new("", data_type.clone(), false);
             check_round_trip_encoding_generated(field, array_gen_provider.copy()).await;
         }
-
-        std::env::remove_var("LANCE_USE_BITPACKING");
     }
 }
