+++ rust/lance-file/src/v2/reader.rs
@@ -970,6 +970,7 @@ pub mod tests {
         encoder::{encode_batch, CoreFieldEncodingStrategy, EncodedBatch},
     };
     use lance_io::stream::RecordBatchStream;
+    use lance_testing::util::EnvVarGuard;
     use log::debug;
 
     use crate::v2::{
@@ -1241,32 +1242,6 @@ pub mod tests {
         )
         .await
         .is_err());
-    }
-
-    struct EnvVarGuard {
-        key: String,
-        original_value: Option<String>,
-    }
-
-    impl EnvVarGuard {
-        fn new(key: &str, new_value: &str) -> Self {
-            let original_value = std::env::var(key).ok();
-            std::env::set_var(key, new_value);
-            Self {
-                key: key.to_string(),
-                original_value,
-            }
-        }
-    }
-
-    impl Drop for EnvVarGuard {
-        fn drop(&mut self) {
-            if let Some(ref value) = self.original_value {
-                std::env::set_var(&self.key, value);
-            } else {
-                std::env::remove_var(&self.key);
-            }
-        }
     }
 
     #[test_log::test(tokio::test)]
