+++ rust/lance-testing/src/lib.rs
@@ -2,3 +2,4 @@
 // SPDX-FileCopyrightText: Copyright The Lance Authors
 
 pub mod datagen;
+pub mod util;
